+++ lib/research_metadata.rb
@@ -2,6 +2,7 @@ require 'puree'
 require 'datacite/mapping'
 require 'research_metadata/transformer/dataset'
 require 'research_metadata/transformer/publication'
+require 'research_metadata/transformer/thesis'
 require 'research_metadata/version'
 
 # Metadata extraction from the Pure Research Information System and